+++ relay-metrics/src/protocol.rs
@@ -164,20 +164,99 @@ fn is_valid_name(name: &str) -> bool {
     false
 }
 
-/// Validates an MRI of the form `<ty>:<ns>/<name>@<unit>`
+/// The namespace of a metric.
 ///
-/// Note that the format used in the statsd protocol is different: Metric names are not prefixed
-/// with `<ty>:` as the type is somewhere else in the protocol.
-pub(crate) fn is_valid_mri(name: &str) -> bool {
-    let mut components = name.splitn(2, ':');
-    if components.next().is_none() {
-        return false;
-    }
+/// Namespaces allow to identify the product entity that the metric got extracted from, and/or
+/// identify the use case that the metric belongs to. These namespaces cannot be defined freely,
+/// instead they are defined by Sentry. Over time, there will be more namespaces as we introduce
+/// new metrics-based products.
+///
+/// Right now this successfully deserializes any kind of string, but in reality only `"sessions"`
+/// (for release health) and `"transactions"` (for metrics-enhanced performance) is supported.
+/// Everything else is dropped both in the metrics aggregator and in the store actor.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum MetricNamespace {
+    /// Metrics extracted from sessions.
+    Sessions,
+    /// Metrics extracted from transaction events.
+    Transactions,
+    /// Metrics that relay either doesn't know or recognize the namespace of, will be dropped before
+    /// aggregating. For instance, an MRI of `c:something_new/foo@none` has the namespace
+    /// `something_new`, but as Relay doesn't support that namespace, it gets deserialized into
+    /// this variant.
+    ///
+    /// Relay currently drops all metrics whose namespace ends up being deserialized as
+    /// `unsupported`. We may revise that in the future.
+    Unsupported,
+}
 
-    if let Some(name_unit) = components.next() {
-        parse_name_unit(name_unit).is_some()
-    } else {
-        false
+impl std::str::FromStr for MetricNamespace {
+    type Err = ParseMetricError;
+
+    fn from_str(ns: &str) -> Result<Self, Self::Err> {
+        match ns {
+            "sessions" => Ok(MetricNamespace::Sessions),
+            "transactions" => Ok(MetricNamespace::Transactions),
+            _ => Ok(MetricNamespace::Unsupported),
+        }
+    }
+}
+
+relay_common::impl_str_serde!(MetricNamespace, "a valid metric namespace");
+
+impl fmt::Display for MetricNamespace {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            MetricNamespace::Sessions => write!(f, "sessions"),
+            MetricNamespace::Transactions => write!(f, "transactions"),
+            MetricNamespace::Unsupported => write!(f, "unsupported"),
+        }
+    }
+}
+
+/// A metric name parsed as MRI, a naming scheme which includes most of the metric's bucket key
+/// (excl. timestamp and tags).
+///
+/// For more information see [`Metric::name`].
+pub struct MetricResourceIdentifier<'a> {
+    /// The metric type.
+    pub ty: MetricType,
+    /// The namespace/usecase for this metric. For example `sessions` or `transactions`. In the
+    /// case of the statsd protocol, a missing namespace is converted into the valueconverted into
+    /// the value `"custom"`.
+    pub namespace: MetricNamespace,
+    /// The actual name, such as `duration` as part of `d:transactions/duration@ms`
+    pub name: &'a str,
+    /// The metric unit.
+    pub unit: MetricUnit,
+}
+
+impl<'a> MetricResourceIdentifier<'a> {
+    /// Parses and validates an MRI of the form `<ty>:<ns>/<name>@<unit>`
+    pub fn parse(name: &'a str) -> Result<Self, ParseMetricError> {
+        let (raw_ty, rest) = name.split_once(':').ok_or(ParseMetricError(()))?;
+        let ty = raw_ty.parse()?;
+
+        let (raw_namespace, rest) = rest.split_once('/').ok_or(ParseMetricError(()))?;
+        let (name, unit) = parse_name_unit(rest).ok_or(ParseMetricError(()))?;
+
+        Ok(Self {
+            ty,
+            namespace: raw_namespace.parse()?,
+            name,
+            unit,
+        })
+    }
+}
+
+impl<'a> fmt::Display for MetricResourceIdentifier<'a> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        // `<ty>:<ns>/<name>@<unit>`
+        write!(
+            f,
+            "{}:{}/{}@{}",
+            self.ty, self.namespace, self.name, self.unit
+        )
     }
 }
 
@@ -206,7 +285,7 @@ pub(crate) fn validate_tag_value(tag_value: &mut String) {
 ///
 /// Returns [`MetricUnit::None`] if no unit is specified. Returns `None` if the name or value are
 /// invalid.
-fn parse_name_unit(string: &str) -> Option<(String, MetricUnit)> {
+fn parse_name_unit(string: &str) -> Option<(&str, MetricUnit)> {
     let mut components = string.split('@');
     let name = components.next()?;
     if !is_valid_name(name) {
@@ -218,7 +297,7 @@ fn parse_name_unit(string: &str) -> Option<(String, MetricUnit)> {
         None => MetricUnit::default(),
     };
 
-    Some((name.to_owned(), unit))
+    Some((name, unit))
 }
 
 /// Hashes the given set value.
@@ -249,10 +328,7 @@ fn parse_value(string: &str, ty: MetricType) -> Option<MetricValue> {
 ///
 /// Returns [`MetricUnit::None`] if no unit is specified. Returns `None` if any of the components is
 /// invalid.
-fn parse_name_unit_value(
-    string: &str,
-    ty: MetricType,
-) -> Option<(String, MetricUnit, MetricValue)> {
+fn parse_name_unit_value(string: &str, ty: MetricType) -> Option<(&str, MetricUnit, MetricValue)> {
     let mut components = string.splitn(2, ':');
     let (name, unit) = components.next().and_then(parse_name_unit)?;
     let value = components.next().and_then(|s| parse_value(s, ty))?;
@@ -349,20 +425,36 @@ fn parse_tags(string: &str) -> Option<BTreeMap<String, String>> {
 
 #[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
 pub struct Metric {
-    /// The name of the metric without its unit.
+    /// The metric resource identifier, in short MRI.
     ///
-    /// Metric names cannot be empty, must start with a letter and can consist of ASCII
-    /// alphanumerics, underscores, slashes, @s and periods.
+    /// MRIs follow three core principles:
+
+    /// 1. **Robustness:** Metrics must be addressed via a stable identifier. During ingestion in
+    ///    Relay and Snuba, metrics are preaggregated and bucketed based on this identifier, so it
+    ///    cannot change over time without breaking bucketing.
+    /// 2. **Uniqueness:** The identifier for metrics must be unique across variations of units and
+    ///    metric types, within and across use cases, as well as between projects and
+    ///    organizations.
+    /// 3. **Abstraction:** The user-facing product changes its terminology over time, and splits
+    ///    concepts into smaller parts. The internal metric identifiers must abstract from that,
+    ///    and offer sufficient granularity to allow for such changes.
+    ///
+    /// MRIs have the format `<type>:<ns>/<name>@<unit>`, comprising the following components:
+    ///
+    /// * **Type:** counter (`c`), set (`s`), distribution (`d`), gauge (`g`), and evaluated (`e`)
+    ///   for derived numeric metrics (the latter is a pure query-time construct and is not relevant
+    ///   to Relay or ingestion). See [`MetricType`].
+    /// * **Namespace:** Identifying the product entity and use case affiliation of the metric. See
+    /// [`MetricNamespace`].
+    /// * **Name:** The display name of the metric in the allowed character set.
+    /// * **Unit:** The verbatim unit name. See [`MetricUnit`].
+    ///
+    /// Parsing a metric (or set of metrics) should not fail hard if the MRI is invalid, so this is
+    /// typed as string. Later in the metrics aggregator, the MRI is parsed using
+    /// [`MetricResourceIdentifier`] and validated for invalid characters as well.
+    /// [`MetricResourceIdentifier`] is also used in the kafka producer to route certain namespaces
+    /// to certain topics.
     pub name: String,
-    /// The unit of the metric value.
-    ///
-    /// Units augment metric values by giving them a magnitude and semantics. There are certain
-    /// types of units that are subdivided in their precision, such as the [`DurationUnit`] for time
-    /// measurements.
-    ///
-    /// The unit can be omitted and defaults to [`MetricUnit::None`].
-    #[serde(default, skip_serializing_if = "MetricUnit::is_none")]
-    pub unit: MetricUnit,
     /// The value of the metric.
     ///
     /// [Distributions](MetricType::Distribution) and [counters](MetricType::Counter) require numeric
@@ -390,39 +482,54 @@ pub struct Metric {
 impl Metric {
     /// Creates a new metric using the MRI naming format.
     ///
+    /// See [`Metric::name`].
+    ///
     /// MRI is the metric resource identifier in the format `<type>:<ns>/<name>@<unit>`. This name
     /// ensures that just the name determines correct bucketing of metrics with name collisions.
     pub fn new_mri(
-        namespace: impl fmt::Display,
-        name: impl fmt::Display,
+        namespace: MetricNamespace,
+        name: impl AsRef<str>,
         unit: MetricUnit,
         value: MetricValue,
         timestamp: UnixTimestamp,
         tags: BTreeMap<String, String>,
     ) -> Self {
         Self {
-            name: format!("{}:{}/{}@{}", value.ty(), namespace, name, unit),
-            unit,
+            name: MetricResourceIdentifier {
+                ty: value.ty(),
+                name: name.as_ref(),
+                namespace,
+                unit,
+            }
+            .to_string(),
             value,
             timestamp,
             tags,
         }
     }
 
+    /// Parse statsd-compatible payload of format
+    /// ```text
+    /// [<ns>/]<name>[@<unit>]:<value>|<type>[|#<tags>]`
+    /// ```
     fn parse_str(string: &str, timestamp: UnixTimestamp) -> Option<Self> {
         let mut components = string.split('|');
 
         let name_value_str = components.next()?;
         let ty = components.next().and_then(|s| s.parse().ok())?;
-        let (name, unit, value) = parse_name_unit_value(name_value_str, ty)?;
+        let (name_and_namespace, unit, value) = parse_name_unit_value(name_value_str, ty)?;
+        let (raw_namespace, name) = name_and_namespace
+            .split_once('/')
+            .unwrap_or(("custom", string));
 
-        let mut metric = Self {
-            name: format!("{}:{}", ty, name),
+        let mut metric = Self::new_mri(
+            raw_namespace.parse().ok()?,
+            name,
             unit,
             value,
             timestamp,
-            tags: BTreeMap::new(),
-        };
+            BTreeMap::new(),
+        );
 
         for component in components {
             if let Some('#') = component.chars().next() {
@@ -543,13 +650,12 @@ mod tests {
 
     #[test]
     fn test_parse_counter() {
-        let s = "foo:42|c";
+        let s = "transactions/foo:42|c";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
         insta::assert_debug_snapshot!(metric, @r###"
         Metric {
-            name: "c:foo",
-            unit: None,
+            name: "c:transactions/foo@none",
             value: Counter(
                 42.0,
             ),
@@ -561,13 +667,12 @@ mod tests {
 
     #[test]
     fn test_parse_distribution() {
-        let s = "foo:17.5|d";
+        let s = "transactions/foo:17.5|d";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
         insta::assert_debug_snapshot!(metric, @r###"
         Metric {
-            name: "d:foo",
-            unit: None,
+            name: "d:transactions/foo@none",
             value: Distribution(
                 17.5,
             ),
@@ -579,7 +684,7 @@ mod tests {
 
     #[test]
     fn test_parse_histogram() {
-        let s = "foo:17.5|h"; // common alias for distribution
+        let s = "transactions/foo:17.5|h"; // common alias for distribution
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
         assert_eq!(metric.value, MetricValue::Distribution(17.5));
@@ -587,13 +692,12 @@ mod tests {
 
     #[test]
     fn test_parse_set() {
-        let s = "foo:e2546e4c-ecd0-43ad-ae27-87960e57a658|s";
+        let s = "transactions/foo:e2546e4c-ecd0-43ad-ae27-87960e57a658|s";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
         insta::assert_debug_snapshot!(metric, @r###"
         Metric {
-            name: "s:foo",
-            unit: None,
+            name: "s:transactions/foo@none",
             value: Set(
                 4267882815,
             ),
@@ -605,13 +709,12 @@ mod tests {
 
     #[test]
     fn test_parse_gauge() {
-        let s = "foo:42|g";
+        let s = "transactions/foo:42|g";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
         insta::assert_debug_snapshot!(metric, @r###"
         Metric {
-            name: "g:foo",
-            unit: None,
+            name: "g:transactions/foo@none",
             value: Gauge(
                 42.0,
             ),
@@ -623,23 +726,25 @@ mod tests {
 
     #[test]
     fn test_parse_unit() {
-        let s = "foo@second:17.5|d";
+        let s = "transactions/foo@second:17.5|d";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
-        assert_eq!(metric.unit, MetricUnit::Duration(DurationUnit::Second));
+        let mri = MetricResourceIdentifier::parse(&metric.name).unwrap();
+        assert_eq!(mri.unit, MetricUnit::Duration(DurationUnit::Second));
     }
 
     #[test]
     fn test_parse_unit_regression() {
-        let s = "foo@s:17.5|d";
+        let s = "transactions/foo@s:17.5|d";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
-        assert_eq!(metric.unit, MetricUnit::Duration(DurationUnit::Second));
+        let mri = MetricResourceIdentifier::parse(&metric.name).unwrap();
+        assert_eq!(mri.unit, MetricUnit::Duration(DurationUnit::Second));
     }
 
     #[test]
     fn test_parse_tags() {
-        let s = "foo:17.5|d|#foo,bar:baz";
+        let s = "transactions/foo:17.5|d|#foo,bar:baz";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Metric::parse(s.as_bytes(), timestamp).unwrap();
         insta::assert_debug_snapshot!(metric.tags, @r###"
@@ -678,7 +783,6 @@ mod tests {
     fn test_serde_json() {
         let json = r#"{
   "name": "foo",
-  "unit": "second",
   "type": "c",
   "value": 42.0,
   "timestamp": 4711,
@@ -692,9 +796,6 @@ mod tests {
         insta::assert_debug_snapshot!(metric, @r###"
         Metric {
             name: "foo",
-            unit: Duration(
-                Second,
-            ),
             value: Counter(
                 42.0,
             ),
@@ -724,7 +825,6 @@ mod tests {
         insta::assert_debug_snapshot!(metric, @r###"
         Metric {
             name: "foo",
-            unit: None,
             value: Counter(
                 42.0,
             ),
@@ -736,7 +836,7 @@ mod tests {
 
     #[test]
     fn test_parse_all() {
-        let s = "foo:42|c\nbar:17|c";
+        let s = "transactions/foo:42|c\nbar:17|c";
         let timestamp = UnixTimestamp::from_secs(4711);
 
         let metrics: Vec<Metric> = Metric::parse_all(s.as_bytes(), timestamp)
@@ -748,7 +848,7 @@ mod tests {
 
     #[test]
     fn test_parse_all_crlf() {
-        let s = "foo:42|c\r\nbar:17|c";
+        let s = "transactions/foo:42|c\r\nbar:17|c";
         let timestamp = UnixTimestamp::from_secs(4711);
 
         let metrics: Vec<Metric> = Metric::parse_all(s.as_bytes(), timestamp)
@@ -760,7 +860,7 @@ mod tests {
 
     #[test]
     fn test_parse_all_empty_lines() {
-        let s = "foo:42|c\n\n\nbar:17|c";
+        let s = "transactions/foo:42|c\n\n\nbar:17|c";
         let timestamp = UnixTimestamp::from_secs(4711);
 
         let metric_count = Metric::parse_all(s.as_bytes(), timestamp).count();
@@ -769,7 +869,7 @@ mod tests {
 
     #[test]
     fn test_parse_all_trailing() {
-        let s = "foo:42|c\nbar:17|c\n";
+        let s = "transactions/foo:42|c\nbar:17|c\n";
         let timestamp = UnixTimestamp::from_secs(4711);
 
         let metric_count = Metric::parse_all(s.as_bytes(), timestamp).count();