+++ tests/integration/test_metrics.py
@@ -46,7 +46,7 @@ def test_metrics(mini_sentry, relay):
     mini_sentry.add_basic_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    metrics_payload = f"foo:42|c\nbar:17|c"
+    metrics_payload = f"transactions/foo:42|c\ntransactions/bar:17|c"
     relay.send_metrics(project_id, metrics_payload, timestamp)
 
     envelope = mini_sentry.captured_events.get(timeout=3)
@@ -61,14 +61,14 @@ def test_metrics(mini_sentry, relay):
         {
             "timestamp": timestamp,
             "width": 1,
-            "name": "c:bar",
+            "name": "c:transactions/bar@none",
             "value": 17.0,
             "type": "c",
         },
         {
             "timestamp": timestamp,
             "width": 1,
-            "name": "c:foo",
+            "name": "c:transactions/foo@none",
             "value": 42.0,
             "type": "c",
         },
@@ -82,7 +82,7 @@ def test_metrics_backdated(mini_sentry, relay):
     mini_sentry.add_basic_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp()) - 24 * 60 * 60
-    metrics_payload = f"foo:42|c"
+    metrics_payload = f"transactions/foo:42|c"
     relay.send_metrics(project_id, metrics_payload, timestamp)
 
     envelope = mini_sentry.captured_events.get(timeout=2)
@@ -96,7 +96,7 @@ def test_metrics_backdated(mini_sentry, relay):
         {
             "timestamp": timestamp,
             "width": 1,
-            "name": "c:foo",
+            "name": "c:transactions/foo@none",
             "value": 42.0,
             "type": "c",
         },
@@ -111,26 +111,24 @@ def test_metrics_with_processing(mini_sentry, relay_with_processing, metrics_con
     mini_sentry.add_full_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    metrics_payload = f"foo:42|c\nbar@second:17|c"
+    metrics_payload = f"transactions/foo:42|c\ntransactions/bar@second:17|c"
     relay.send_metrics(project_id, metrics_payload, timestamp)
 
     metrics = metrics_by_name(metrics_consumer, 2)
 
-    assert metrics["c:foo"] == {
+    assert metrics["c:transactions/foo@none"] == {
         "org_id": 1,
         "project_id": project_id,
-        "name": "c:foo",
-        "unit": "none",
+        "name": "c:transactions/foo@none",
         "value": 42.0,
         "type": "c",
         "timestamp": timestamp,
     }
 
-    assert metrics["c:bar"] == {
+    assert metrics["c:transactions/bar@second"] == {
         "org_id": 1,
         "project_id": project_id,
-        "name": "c:bar",
-        "unit": "second",
+        "name": "c:transactions/bar@second",
         "value": 17.0,
         "type": "c",
         "timestamp": timestamp,
@@ -157,18 +155,17 @@ def test_metrics_full(mini_sentry, relay, relay_with_processing, metrics_consume
 
     # Send two events to downstream and one to upstream
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    downstream.send_metrics(project_id, f"foo:7|c", timestamp)
-    downstream.send_metrics(project_id, f"foo:5|c", timestamp)
+    downstream.send_metrics(project_id, f"transactions/foo:7|c", timestamp)
+    downstream.send_metrics(project_id, f"transactions/foo:5|c", timestamp)
 
-    upstream.send_metrics(project_id, f"foo:3|c", timestamp)
+    upstream.send_metrics(project_id, f"transactions/foo:3|c", timestamp)
 
     metric = metrics_consumer.get_metric(timeout=6)
     metric.pop("timestamp")
     assert metric == {
         "org_id": 1,
         "project_id": project_id,
-        "name": "c:foo",
-        "unit": "none",
+        "name": "c:transactions/foo@none",
         "value": 15.0,
         "type": "c",
     }
@@ -272,7 +269,6 @@ def test_session_metrics_non_processing(
                 "timestamp": ts,
                 "width": 1,
                 "type": "d",
-                "unit": "second",
                 "value": [1947.49],
             },
             {
@@ -387,7 +383,6 @@ def test_session_metrics_processing(
         "timestamp": expected_timestamp,
         "name": "c:sessions/session@none",
         "type": "c",
-        "unit": "none",
         "value": 1.0,
         "tags": {
             "sdk": "raven-node/2.6.3",
@@ -403,7 +398,6 @@ def test_session_metrics_processing(
         "timestamp": expected_timestamp,
         "name": "s:sessions/user@none",
         "type": "s",
-        "unit": "none",
         "value": [1617781333],
         "tags": {
             "sdk": "raven-node/2.6.3",
@@ -419,7 +413,6 @@ def test_session_metrics_processing(
         "timestamp": expected_timestamp,
         "name": "d:sessions/duration@second",
         "type": "d",
-        "unit": "second",
         "value": [1947.49],
         "tags": {
             "sdk": "raven-node/2.6.3",
@@ -544,7 +537,6 @@ def test_transaction_metrics(
         **common,
         "name": "d:transactions/measurements.foo@none",
         "type": "d",
-        "unit": "none",
         "value": [1.2, 2.2],
     }
 
@@ -552,7 +544,6 @@ def test_transaction_metrics(
         **common,
         "name": "d:transactions/measurements.bar@none",
         "type": "d",
-        "unit": "none",
         "value": [1.3],
     }
 
@@ -562,7 +553,6 @@ def test_transaction_metrics(
         **common,
         "name": "d:transactions/breakdowns.span_ops.ops.react.mount@millisecond",
         "type": "d",
-        "unit": "millisecond",
         "value": [9.910106, 9.910106],
     }
 
@@ -570,7 +560,6 @@ def test_transaction_metrics(
         **common,
         "name": "d:transactions/breakdowns.span_ops.total.time@millisecond",
         "type": "d",
-        "unit": "millisecond",
         "value": [9.910106, 9.910106],
     }
 
@@ -595,17 +584,17 @@ def test_graceful_shutdown(mini_sentry, relay):
 
     # Backdated metric will be flushed immediately due to debounce delay
     past_timestamp = timestamp - 1000
-    metrics_payload = f"foo:42|c"
+    metrics_payload = f"transactions/foo:42|c"
     relay.send_metrics(project_id, metrics_payload, past_timestamp)
 
     # Future timestamp will not be flushed regularly, only through force flush
-    metrics_payload = f"bar:17|c"
+    metrics_payload = f"transactions/bar:17|c"
     future_timestamp = timestamp + 60
     relay.send_metrics(project_id, metrics_payload, future_timestamp)
     relay.shutdown(sig=signal.SIGTERM)
 
     # Try to send another metric (will be rejected)
-    metrics_payload = f"zap:666|c"
+    metrics_payload = f"transactions/zap:666|c"
     with pytest.raises(requests.ConnectionError):
         relay.send_metrics(project_id, metrics_payload, timestamp)
 
@@ -619,14 +608,14 @@ def test_graceful_shutdown(mini_sentry, relay):
         {
             "timestamp": future_timestamp,
             "width": 1,
-            "name": "c:bar",
+            "name": "c:transactions/bar@none",
             "value": 17.0,
             "type": "c",
         },
         {
             "timestamp": past_timestamp,
             "width": 1,
-            "name": "c:foo",
+            "name": "c:transactions/foo@none",
             "value": 42.0,
             "type": "c",
         },